hospital ManagmentSystem and addede redirectUrl

+++ SpringSecurity/src/main/java/com/org/SpringSecurity/dto/LoginresponseDto.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class LoginresponseDto {
     String jwt;
-    int userId;
+    Long userId;
 }
